feat: enable mypy --strict for decision/, engine/, event/, features/, execution/, alpha/, attribution/, runner/backtest — zero errors
Extends strict type checking from core/regime/state (previous commit) to 7 more
directories (70 files, 160+ checked). Fixes include: generic type parameters,
return type annotations, Any-to-concrete casts for Rust FFI calls, None guards,
and unused import cleanup.

+++ attribution/live_tracker.py
@@ -203,7 +203,7 @@ class LiveSignalTracker:
             return 0.0
 
         # Annualize (hourly -> yearly: sqrt(8760))
-        return (mean_ret / std) * (8760 ** 0.5)
+        return float((mean_ret / std) * (8760 ** 0.5))
 
     def compute_weight_recommendations(
         self,
